updated IntegrationPointStub.kt
